refactor(cardgroups): extract useDeleteCardgroup hook from cardgroup-header
